feat: usando type Half e melhorando consumo de RAM, e organizando codigo

+++ transactions/models/ReferenceDatasetItem.cs
@@ -5,7 +5,7 @@ namespace rinha_backend_csharp_2026.transactions.models
     public sealed class ReferenceDatasetItem
     {
         [JsonPropertyName("vector")]
-        public required float[] Vector { get; init; }
+        public required Half[] Vector { get; init; }
 
         [JsonPropertyName("label")]
         public required string Label { get; init; }
